refactor routine: 패키지 정리
