+++ src/main/java/com/groomthon/habiglow/domain/daily/service/DailyRoutineService.java
@@ -66,9 +66,9 @@ public class DailyRoutineService {
         
         // FULL_SUCCESS인 경우 증가, 아니면 리셋
         if (performanceLevel == PerformanceLevel.FULL_SUCCESS) {
-            routine.updateGrowthConfiguration(routine.getGrowthConfiguration().incrementCurrentCycleDays());
+            routine.updateGrowthConfiguration(routine.getGrowthConfiguration().withIncrementedCycle());
         } else {
-            routine.updateGrowthConfiguration(routine.getGrowthConfiguration().resetCurrentCycleDays());
+            routine.updateGrowthConfiguration(routine.getGrowthConfiguration().withResetCycle());
         }
     }
 }
