+++ src/main/java/com/groomthon/habiglow/domain/member/dto/request/UpdateMemberInterestsRequest.java
@@ -2,7 +2,7 @@ package com.groomthon.habiglow.domain.member.dto.request;
 
 import java.util.List;
 
-import com.groomthon.habiglow.domain.routine.entity.RoutineCategory;
+import com.groomthon.habiglow.domain.routine.common.RoutineCategory;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
