+++ src/main/java/com/groomthon/habiglow/domain/member/entity/MemberEntity.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import com.groomthon.habiglow.domain.routine.entity.RoutineCategory;
+import com.groomthon.habiglow.domain.routine.common.RoutineCategory;
 import com.groomthon.habiglow.global.entity.BaseTimeEntity;
 import com.groomthon.habiglow.global.oauth2.entity.SocialType;
