+++ src/main/java/com/groomthon/habiglow/domain/routine/dto/request/CreateRoutineRequest.java
@@ -1,7 +1,7 @@
 package com.groomthon.habiglow.domain.routine.dto.request;
 
-import com.groomthon.habiglow.domain.routine.entity.RoutineCategory;
-import com.groomthon.habiglow.domain.routine.entity.TargetType;
+import com.groomthon.habiglow.domain.routine.common.RoutineCategory;
+import com.groomthon.habiglow.domain.routine.common.TargetType;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
