+++ src/main/java/com/groomthon/habiglow/domain/routine/dto/response/RoutineResponse.java
@@ -2,9 +2,9 @@ package com.groomthon.habiglow.domain.routine.dto.response;
 
 import java.time.LocalDateTime;
 
-import com.groomthon.habiglow.domain.routine.entity.RoutineCategory;
+import com.groomthon.habiglow.domain.routine.common.RoutineCategory;
 import com.groomthon.habiglow.domain.routine.entity.RoutineEntity;
-import com.groomthon.habiglow.domain.routine.entity.TargetType;
+import com.groomthon.habiglow.domain.routine.common.TargetType;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
