+++ src/main/java/com/groomthon/habiglow/domain/routine/dto/response/adaptation/GrowthReadyRoutineResponse.java
@@ -1,11 +1,11 @@
-package com.groomthon.habiglow.domain.routine.dto.response;
+package com.groomthon.habiglow.domain.routine.dto.response.adaptation;
 
 import java.time.LocalDate;
 
 import com.groomthon.habiglow.domain.daily.entity.DailyRoutineEntity;
-import com.groomthon.habiglow.domain.routine.entity.RoutineCategory;
+import com.groomthon.habiglow.domain.routine.common.RoutineCategory;
 import com.groomthon.habiglow.domain.routine.entity.RoutineEntity;
-import com.groomthon.habiglow.domain.routine.entity.TargetType;
+import com.groomthon.habiglow.domain.routine.common.TargetType;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
