+++ src/main/java/com/groomthon/habiglow/domain/routine/dto/response/adaptation/RoutineAdaptationCheckResponse.java
@@ -1,4 +1,4 @@
-package com.groomthon.habiglow.domain.routine.dto.response;
+package com.groomthon.habiglow.domain.routine.dto.response.adaptation;
 
 import java.util.List;
