+++ src/main/java/com/groomthon/habiglow/domain/routine/service/ReductionAnalysisService.java
@@ -10,8 +10,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.groomthon.habiglow.domain.daily.entity.DailyRoutineEntity;
 import com.groomthon.habiglow.domain.daily.repository.DailyRoutineRepository;
-import com.groomthon.habiglow.domain.routine.dto.response.RoutineAdaptationCheckResponse;
-import com.groomthon.habiglow.domain.routine.dto.response.ReductionReadyRoutineResponse;
+import com.groomthon.habiglow.domain.routine.dto.response.adaptation.RoutineAdaptationCheckResponse;
+import com.groomthon.habiglow.domain.routine.dto.response.adaptation.ReductionReadyRoutineResponse;
 import com.groomthon.habiglow.domain.routine.entity.RoutineEntity;
 import com.groomthon.habiglow.domain.routine.repository.RoutineRepository;
 import com.groomthon.habiglow.domain.routine.service.strategy.ReductionStrategy;
